# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Grafana API",
        "description": "API documentation for Grafana integration",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/api/getDashboards": {
            "get": {
                "parameters": [
                    {
                        "name": "grafana_base_url",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API URL"
                    },
                    {
                        "name": "grafana_token",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API Token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of dashboards"
                    }
                }
            }
        },
        "/api/getDashboard/{id}": {
            "get": {
                "parameters": [
                    {
                        "name": "grafana_base_url",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API URL"
                    },
                    {
                        "name": "grafana_token",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API Token"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "Id of dashboard to fetch"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard details"
                    }
                }
            }
        },
        "/api/getDatasources": {
            "get": {
                "parameters": [
                    {
                        "name": "grafana_base_url",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API URL"
                    },
                    {
                        "name": "grafana_token",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API Token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of datasources"
                    }
                }
            }
        },
        "/api/getDatasource/{id}": {
            "get": {
                "parameters": [
                    {
                        "name": "grafana_base_url",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API URL"
                    },
                    {
                        "name": "grafana_token",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "description": "Grafana API Token"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "type": "string",
                        "required": true,
                        "description": "Id of datasource to fetch"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datasource details"
                    }
                }
            }
        }
    }
}